Move Firestore code under data package
