+++ CovidWatch/app/src/main/java/org/covidwatch/android/data/firestore/LocalContactEventsViewModel.kt
@@ -1,4 +1,4 @@
-package org.covidwatch.android.firestore
+package org.covidwatch.android.data.firestore
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
